+++ GEMM-FP64/src/kernels/1_dgemm_basic.cuh
@@ -14,10 +14,10 @@
 /// @param B Pointer to B matrix (K x N)
 /// @param C Pointer to C matrix (M x N)
 template<unsigned int TS>
-__global__ void dgemm_basic(double alpha, double beta, int M, int N, int K, double* A, double* B, double* C) {
-  extern __shared__ double sm[];
-  double* sA = &sm[0];
-  double* sB = &sm[TS * TS];
+__global__ void dgemm_basic(float alpha, float beta, int M, int N, int K, float* A, float* B, float* C) {
+  extern __shared__ float sm[];
+  float* sA = &sm[0];
+  float* sB = &sm[TS * TS];
 
   const unsigned int tx = threadIdx.x;
   const unsigned int ty = threadIdx.y;
@@ -25,7 +25,7 @@ __global__ void dgemm_basic(double alpha, double beta, int M, int N, int K, doub
   unsigned int bm = blockIdx.y * TS;
   unsigned int bn = blockIdx.x * TS;
 
-  double acc_reg = 0.0;
+  float acc_reg = 0.0;
   for(unsigned int bk = 0; bk < K; bk += TS) {
     sA[ty * TS + tx] = A[(bm + ty) * K + (bk + tx)];
     sB[ty * TS + tx] = B[(bk + ty) * N + (bn + tx)];
